sys/net: fix type-limits in sc_netif

+++ sys/shell/commands/sc_netif.c
@@ -552,7 +552,7 @@ static void _netif_list(kernel_pid_t dev)
 
 static int _netif_set_u16(kernel_pid_t dev, netopt_t opt, char *u16_str)
 {
-    unsigned int res;
+    unsigned long int res;
     bool hex = false;
 
     if (_is_number(u16_str)) {
